selecting contactListItem stores contact id in redux, contact id can be selected from global store
