+++ src/Components/ContactBook.js
@@ -1,12 +1,10 @@
 import React from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { selectSelectedContactId } from '../store/ContactBook/selectors'
 import { makeStyles } from '@material-ui/core/styles'
-import {
-	Paper,
-	Grid,
-	Typography,
-	List,
-} from '@material-ui/core'
+import { Paper, Grid, List } from '@material-ui/core'
 import ContactListItem from './ContactListItem'
+import ContactCard from './ContactCard'
 
 const useStyles = makeStyles((theme) => ({
 	contactList: {
@@ -19,20 +17,31 @@ const useStyles = makeStyles((theme) => ({
 export default function ContactBook(props) {
 	const classes = useStyles()
 	const contacts = props.contacts
-
+	const selectedContactId = useSelector(selectSelectedContactId)
+	const selectedContact = contacts.find(
+		(contact) => contact.id === selectedContactId
+    )
+    
 	return (
 		<Grid container direction='row' justify='flex-start'>
-			<Grid item xs={6}>
-				<Paper>
+			<Paper>
+				<Grid item xs={6}>
 					<List className={classes.contactList}>
 						{contacts.map((contact) => {
 							return (
-								<ContactListItem contact={contact}/>
+								<ContactListItem
+									key={contact.id}
+									index={contact.id}
+									contact={contact}
+								/>
 							)
 						})}
 					</List>
-				</Paper>
-			</Grid>
+				</Grid>
+				<Grid item>
+					{selectedContact ? <ContactCard contact={selectedContact} /> : null}
+				</Grid>
+			</Paper>
 		</Grid>
 	)
 }
