+++ src/store/ContactBook/actions.js
@@ -1,0 +1,6 @@
+export function setSelectedContactId(data) {
+	return {
+		type: 'STORE_SELECTED_CONTACT',
+		payload: data,
+	}
+}
