+++ src/store/ContactBook/selectors.js
@@ -1,0 +1,3 @@
+export const selectSelectedContactId = (state) => {
+    return state.ContactBook.selectedContactId
+}
