+++ src/store/rootReducer.js
@@ -1,10 +1,12 @@
-import { combineReducers } from "redux";
-import appState from "./appState/reducer";
-import user from "./user/reducer";
-import appFeed from "./appFeed/reducer"
+import { combineReducers } from 'redux'
+import appState from './appState/reducer'
+import user from './user/reducer'
+import appFeed from './appFeed/reducer'
+import ContactBook from './ContactBook/reducer'
 
 export default combineReducers({
-  appFeed,
-  appState,
-  user
-});
+	ContactBook,
+	appFeed,
+	appState,
+	user,
+})
